+++ src/lib/isMac.js
@@ -1,3 +1,5 @@
+import getOS from './getOS';
+
 /**
  * 检查是不是Mac环境。
  * @returns {boolean} 真/假
@@ -12,8 +14,7 @@
  *
  */
 function isMac() {
-  let p = navigator.platform;
-  return p === 'Mac68K' || p === 'MacPPC' || p === 'Macintosh' || p === 'MacIntel';
+  return getOS().name === 'Mac OS';
 }
 
 export default isMac;