+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # GHANGELOG
 
+## 2.0.6
+
+- 优化了defineEnum()，增加key；
+- 新增了stringToBase64、base64ToString方法；
+
 ## 2.0.5
 
 - 优化getQueryString；